api/pyproject.toml
[project]
name = "punya-api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datamodel-code-generator[http]>=0.35.0",
    "fastapi[standard]>=0.120.2",
    "geoalchemy2>=0.18.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pyzmq>=27.1.0",
    "shapely>=2.1.2",
    "sqlmodel>=0.0.27",
    "taskiq>=0.11.19",
    "taskiq-fastapi>=0.3.5",
    "taskiq-redis>=1.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]

[project.scripts]
eddi-models = 'commands:sync_eddi_models'

[tool.ruff]
select = ["I", "FURB", "TRY"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
